Move HuggingFace extractor helpers to module
