+++ neuralset-repo/neuralset/extractors/image.py
@@ -17,6 +17,7 @@ from neuralset import base, utils
 from neuralset.events import etypes
 
 from . import base as extractor_base
+from . import huggingface
 
 logger = logging.getLogger(__name__)
 CLUSTER_DEFAULTS: dict[str, tp.Any] = dict(
@@ -99,7 +100,7 @@ def _huggingface_image_event_uid(event: etypes.Image | etypes.Video) -> str:
     return str(event.study_relative_path())
 
 
-class HuggingFaceImageConfig(extractor_base.HuggingFaceConfig):
+class HuggingFaceImageConfig(huggingface.HuggingFaceConfig):
     processor_kwargs: dict[str, tp.Any] | None = {"do_rescale": False}
     HF_CLASS_DEFAULTS: tp.ClassVar[dict[str, dict[str, str]]] = {
         "clip": {
@@ -113,7 +114,7 @@ class HuggingFaceImageConfig(extractor_base.HuggingFaceConfig):
     }
 
 
-class HuggingFaceImage(extractor_base.BaseStatic, extractor_base.HuggingFaceMixin):
+class HuggingFaceImage(extractor_base.BaseStatic, huggingface.HuggingFaceMixin):
     """Compute image embeddings using transformer-based models obtained through HuggingFace API.
 
     Parameters
@@ -143,13 +144,13 @@ class HuggingFaceImage(extractor_base.BaseStatic, extractor_base.HuggingFaceMixi
         return (
             ["batch_size"]
             + extractor_base.BaseStatic._exclude_from_cls_uid()
-            + extractor_base.HuggingFaceMixin._exclude_from_cls_uid()
+            + huggingface.HuggingFaceMixin._exclude_from_cls_uid()
         )
 
     def _exclude_from_cache_uid(self) -> list[str]:
         return extractor_base.BaseStatic._exclude_from_cache_uid(
             self
-        ) + extractor_base.HuggingFaceMixin._exclude_from_cache_uid(self)
+        ) + huggingface.HuggingFaceMixin._exclude_from_cache_uid(self)
 
     def _iter_image_latents(
         self, events: tp.Sequence[etypes.Image], aggregate_layers: bool
