+++ neuralset-repo/neuralset/extractors/text.py
@@ -21,7 +21,8 @@ import neuralset as ns
 from neuralset import events as _ev  # avoid circular import
 from neuralset import utils
 from neuralset.base import TimedArray
-from neuralset.extractors.base import BaseStatic, HuggingFaceConfig, HuggingFaceMixin
+from neuralset.extractors.base import BaseStatic
+from neuralset.extractors.huggingface import HuggingFaceConfig, HuggingFaceMixin
 
 # pylint: disable=attribute-defined-outside-init
 # pylint: disable=unused-variable
